Revert "separate api functions to different files, set up config and images endpoint"
This reverts commit a183172bb5968bd997370750d0ce61591b68b6c6.
