--- src/api/images.tsx
@@ -1,13 +1,0 @@
-import catApiClient from "./api";
-import { QueryParams } from "@/types/common";
-
-export const getImages = async (queryParams: QueryParams = {}) => {
-  try {
-    const { data } = await catApiClient.get("/images/search", {
-      params: queryParams,
-    });
-    return data;
-  } catch (err) {
-    console.log(err);
-  }
-};
