--- src/types/common.tsx
@@ -1,1 +1,0 @@
-export type QueryParams = Record<string, string>;
